Add initial implementation of PasswordVault class
I have introduced the initial implementation of the PasswordVault class,
which store and retrieves passwords using a JSON file as the storage
medium. Currently, there is no encryption mechanism, however one will be
added at a later date.

The write_password method allows users to store passwords for a given
service, while the get_password method retrieves the password associated
with a specified service.

This sets the foundation for further updates to the password manager
functionality.