Add del command & encapsulate code for organization
I have added a del command to delete a specified service and its password.
The del command is configured to the PasswordVault's del_password
method.

Secondly, I have encapsulated the lines of code that validated the
attempted master password into a method for the PasswordVault class
called _validate_master_password. I did this to reduce the redundancy
of those lines appearing multiple times.

In addition, I have encapsulated multiple lines of code in the App class
that configured commands into class methods. This was done
for the organization.

Lastly, I have updated the README.md file to include information on the
new del command.