Refactor storage mechanism to binary file format
I made this change to address issues with the storage of encrypted data
and hashes in a JSON file format. In order to resolve these issues and
simplify the process, the storage mechanism has been refactored to use
a binary file format instead.

By transitioning to a binary file, we can directly store our encrypted
data as bytes, eliminating the need to decode the encrypted data into a
string using various character encodings. This simplifies our code and
streamlines the process.