Print data in a more readable format
I have changed the data format printed by the print_data method to be
more readable. Before, data printed was messy, and it wasn't easy to
discern which encryption belonged to what service. Now service/password
pairs are printed on new lines.

In addition, if a data file does not exist, the method will print
that instead.