Change several method names in PasswordVault class
I have made the following changes to the method names for clarity of
purpose:

- `_validate_master_password` -> `_verify_master_password`

I made this change to emphasize that the method verifies whether the
inputted master password matches the one saved. Using "verify" better
reflects its role in checking the correctness of the password, as
opposed to "validate," which might imply pre-saving checks handled by the
validator class.

- `_get_validated_password` -> `_wait_for_validated_password`

This method will continuously prompt the user to enter a valid password.
The execution of the code will wait until a valid password is provided,
making the name "wait_for_validated_password" more suitable to convey its
functionality.

- `_get_hashed_master_pass` -> `_load_master_pass_from_data`
- `_get_stored_salt` -> `_load_salt_from_data`

I made these changes to align the naming with the convention used by
`_load_data_from_file`. The verb "load" better represents the action
of accessing the master password and salt from an external source, while
"get" is typically used for retrieving data members from classes.

These updated method names aim to enhance the clarity and maintainability
of the PasswordVault class, making it easier to understand the purpose
and behavior of each method.